also test for faccessat
git-svn-id: https://godirepo.camlcity.org/svn/lib-ocamlnet2/trunk/code@1516 96e5b9e0-aa12-0410-9e2e-b23d24e56262

+++ src/netsys/configtests/atfunctions.c
@@ -1,5 +1,4 @@
 #define _GNU_SOURCE
-#define _XOPEN_SOURCE 700
 #define _ATFILE_SOURCE
 
 #include <errno.h>
@@ -13,6 +12,11 @@
 
 /* reminder: we return the exit code, and 0 means success */
 
+/* Solaris currently has openat and unlinkat but not the other
+ * at functions. We want them as a whole only - so test here
+ * explicitly for one of the other functions, faccessat
+ */
+
 value check(value dummy) {
     int fd1, fd2;
 
@@ -20,6 +24,7 @@ value check(value dummy) {
     if (fd1 == -1) return Val_int(1);
     fd2 = openat(fd1, "atfiletest", O_RDWR|O_CREAT, 0600);
     if (fd2 == -1) return Val_int(1);
+    if (faccessat(fd1, "atfiletest", F_OK, 0) == -1) return Val_int(1);
     if (unlinkat(fd1, "atfiletest", 0) == -1) return Val_int(1);
     return Val_int(0);
 }
